pyproject.toml
[project]
name = "secondworld"
version = "0.1.0"
description = "Multi-agent novel drafting CLI."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "langchain",
  "langchain-deepseek",
  "langchain-openai",
  "langgraph",
  "pyyaml",
  "python-dotenv",
]

[dependency-groups]
dev = [
  "pytest",
]
